+++ fiddles/bash/fiddle-0004-CurlDownload/script.sh
@@ -3,4 +3,4 @@
 clear;
 echo "$0" | sed 's/\.\///g' | awk '{print toupper($0)}'
 
-curl https://www.jetbrains.com/idea/download/download-thanks.html?platform=mac -o intelliJ.zip
+curl http://apache.claz.org/logging/log4j/1.2.17/log4j-1.2.17.zip -o log4j-1.2.17.zip