+++ scripts/bin/setup/_setup.sh
@@ -26,6 +26,10 @@ function setup() {
                 source bin/setup/mac/_brew.sh;
                 install || exit $?;
                 ;;
+              'gh')
+                source bin/setup/mac/_gh.sh;
+                install || exit $?;
+                ;;
               'joe')
                 source bin/setup/mac/_joe.sh;
                 install || exit $?;
@@ -36,6 +40,10 @@ function setup() {
                 ;;
               'live-server')
                 source bin/setup/mac/_live-server.sh;
+                install || exit $?;
+                ;;
+              'nativescript')
+                source bin/setup/mac/_nativescript.sh;
                 install || exit $?;
                 ;;
               'zsh')
@@ -52,11 +60,13 @@ function setup() {
                 ;;
               'all')
                 setup ${_os} "brew" || exit $?;
+                setup ${_os} "gh" || exit $?;
                 setup ${_os} "node" || exit $?;
                 setup ${_os} "nvm" || exit $?;
                 setup ${_os} "joe" || exit $?;
                 setup ${_os} "js-beautify" || exit $?;
                 setup ${_os} "live-server" || exit $?;
+                setup ${_os} "nativescript" || exit $?;
                 setup ${_os} "zsh" || exit $?;
                 ;;
                *) exit 86;